pyproject.toml
[tool.poetry]
name = "auth-services"
version = "0.1.0"
description = "Service responsible to register and authenticate users"
authors = ["Almerindo Uazela <almerindo.uazela@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["all"], version = "^0.115.6"}
uvicorn = {extras = ["standard"], version = "^0.34.0"}
sqlalchemy = "^2.0.36"
mysql-connector-python = "^9.1.0"
redis = "^5.2.1"
aioredis = "^2.0.1"
pydantic = "^2.0"
python-dotenv = "^1.0.1"
alembic = "^1.14.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
httpx = "^0.28.1"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.1"
pydantic-settings = "^2.0"
bcrypt = "3.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"